refactor(nextly): move field-group table provisioning beside its schema service
Runtime registration and companion provisioning were private functions inside
the components dispatcher, which is why the two other create transports could
not reach the schema half and shipped a registry row describing a table that
was never created.

Both bodies move unchanged. Verified byte-identical against main; the only
difference is the depth of three dynamic-import paths, which the new location
changes.
